wood-steel-vbackend/requirements.txt
flask
flask-cors
pymongo[srv]
python-dotenv
flask-jwt-extended
bcrypt
gunicorn
pandas
openpyxl
